Changing a category takes two parameters.
Category change can also take in a photo to change to upon changing the
category. Before a photo and category were changed in parallel. I
believe this may have read to the odd race condition that would have
the incorrect photo show on the first category change a user does in
the page. I've also moved more functionality from the router into the
controller where it belongs.