Doing the Graph API the right way.
I failed to realize you need an App ID and Secret token to use the
Graph api. Simply making get requests to the API results in periodic
termination of get requests which was causing outages in photo delivery
with high frequency. Using the Facebook PHP SDK instead of a lame
'file_get_contents' call is the way to go.