Adds JPG minification
grunt-contrib-imagemin doesn’t compress as well as TinyPNG, though it
nearly does on the tiny dimension images (which to me is the most
important since those are being embedded directly into HTML and not
lazy-loaded). But the trade off in size is worth the convenience of
automation in Grunt with a no-cost package.